Completed Streaming for awhile
The map is currently in World Composition mode. This is the streaming
mode Molly is using. For now the other three modes (Level streaming
modes) have been investigated. The three modes are Level Streaming
Volumes, Blueprint Level Streaming, and C++ Level Streaming. All three
of these methods rely on having defined volumes that activate the
desired level as long as the player is inside the volume. When the
player leaves the volume the level is unloaded after a short delay.