Removed Drag and adjusted max thrust for the engine.
Added all the starter pack that I used since I am distributing this I
wanted it to be complete.
I still need to fix the thrust model. The aircraft is at the correct
weight for a spitfire and full throttle in level flight it makes about
320 kts which is correct. I recommend flying it at 120 knots (more or
less) since at 320 it is easy to fly out of the world box. If that
happens hit escape and start again. Gravity is enabled so when the
aircraft climbs it will slow down and when it dives it will speed up.
The real spitfire had a max climb rate of about 2200 fpm but this
aircraft will do way better since it has no drag. Likewise when it dives
there is no restriction. The best strategy is to do loops followed by
half rolls to stay in the box. The center of the box is the little land
island so keeping this in site is a good idea.